pyproject.toml
[project]
name = "bubseek"
version = "0.0.1"
description = "Intelligent Perception and Explainable Analytics in Collaboration ｜ Bub & seekdb Inside"
authors = [{ name = "Chojan Shang", email = "psiace@apache.org" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.12,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "bub",
    # pyobvector imports Expression from sqlglot; sqlglot 30+ removed that export.
    "sqlglot>=26.0.0,<30.0.0",
    "republic>=0.5.4",
    "typer>=0.12.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "bub-feishu",
    "bub-dingtalk",
    "bub-wechat",
    "bub-discord",
    "bub-web-search",
    "bub-tapestore-sqlalchemy",
    "pyobvector>=0.2.22",
    "bubseek-schedule",
    "bubseek-marimo",
]

[project.urls]
Repository = "https://github.com/ob-labs/bubseek"

[project.entry-points."bub"]
oceanbase-dialect = "bubseek.oceanbase:register"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=4.5.1",
    "tox-uv>=1.29.0",

    "ty>=0.0.14",

    "pandas",

    "ruff>=0.14.14",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.2",
]

[build-system]
requires = ["pdm-backend", "pdm-build-skills>=0.1.0a3"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = [
    "src/bubseek",
    "src/skills",
]
skills = [
    { git = "https://github.com/PsiACE/skills.git", subpath = "skills", include = ["friendly-python", "piglet"] },
    { git = "https://github.com/bubbuild/bub-contrib.git", subpath = ".agents/skills", include = ["plugin-creator"] },
]

# Lock file is always resolved against PyPI. For faster install during dev, use:
#   UV_INDEX_URL=https://pypi.tuna.tsinghua.edu.cn/simple uv sync
[tool.uv]
index-url = "https://pypi.org/simple"

[tool.uv.sources]
# bub-contrib: branch=main so bundled runtime packages (e.g. bub-wechat, bub-discord) exist at resolve time; uv.lock pins the exact commit.
bub = { git = "https://github.com/bubbuild/bub.git" }
bub-dingtalk = { git = "https://github.com/bubbuild/bub-contrib.git", branch = "main", subdirectory = "packages/bub-dingtalk" }
bub-feishu = { git = "https://github.com/bubbuild/bub-contrib.git", branch = "main", subdirectory = "packages/bub-feishu" }
bub-wechat = { git = "https://github.com/bubbuild/bub-contrib.git", branch = "main", subdirectory = "packages/bub-wechat" }
bub-discord = { git = "https://github.com/bubbuild/bub-contrib.git", branch = "main", subdirectory = "packages/bub-discord" }
bub-web-search = { git = "https://github.com/bubbuild/bub-contrib.git", branch = "main", subdirectory = "packages/bub-web-search" }
bub-tapestore-sqlalchemy = { git = "https://github.com/bubbuild/bub-contrib.git", branch = "main", subdirectory = "packages/bub-tapestore-sqlalchemy" }
bubseek-schedule = { workspace = true }
bubseek-marimo = { workspace = true }

[tool.uv.workspace]
members = [
    "contrib/bubseek-schedule",
    "contrib/bubseek-marimo",
]

[tool.ty.environment]
python = "./.venv"
python-version = "3.12"

[tool.ty.src]
# references/ = third-party or example code. insights/ = PEP 723 scripts (marimo, pandas, etc.)
# whose deps are declared in-script only; CI does not install them for type-checking.
exclude = ["references", "insights", "scripts/sitecustomize.py"]

[tool.pytest.ini_options]
testpaths = ["tests", "contrib/bubseek-schedule/src/tests", "contrib/bubseek-marimo/tests"]

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"contrib/*/src/tests/*" = ["S101"]
"contrib/*/tests/*" = ["S101"]
"contrib/*/src/**/*.py" = ["TRY003"]
"src/skills/github-repo-cards/scripts/*.py" = ["S603", "S607"]

[tool.ruff.format]
preview = true
